requirements.txt
fastapi==0.68.0 # Web framework used for building the API
uvicorn==0.15.0 # ASGI server for running FastAPI app
jinja2==3.0.1 # Template engine for rendering HTML
aiofiles==0.7.0 # Asynchronous file operations
python-multipart==0.0.5 # Parser for Python
passlib==1.7.4 # Password hashing library
pydantic==1.8.2 # Data validation
bcrypt==3.2.0 # Password hashing algorithm
python-jose==3.3.0 # JSON Web Token implementation
cryptography==43.0.1 # Cryptography!
pylint==3.3.0 # Python code linter
pytest==7.4.2 # Testing framework
pytest-cov==4.1.0 # Coverage plugin for pytest
requests==2.31.0 # HTTP library for testing API endpoints